feat: add layout component to maintain same layout to all pages, move navigation bar code to file
